projects/finops/py_requirements.txt
requests
datetime
pandas
azure-storage-blob
python-decouple